게시물 삭제는 해당 게시물을 작성한 사람만 지울 수 있도록

+++ server/src/Post/postController.js
@@ -28,10 +28,10 @@ exports.deletePost = async (req, res) => {
 
   const postUserIdx = await postProvider.getPostUserIdx(postIdx);
   //특정 postIdx의 userIdx와 동일할 때에만 해당 게시물을 삭제할 수 있도록 하여야 한다.
-
   if (!postIdx)
     return res.send(basicResponse(baseResponseStatus.PARAMS_NOT_EXACT));
-
+  if (userIdx != postUserIdx)
+    return res.send(basicResponse(baseResponseStatus.USER_NOT_AUTH));
   const result = await postService.deletePost(postIdx);
 
   return res.send(result);
